small compatibility change in clipping-system
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@817 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CBytestream.cpp
@@ -417,7 +417,7 @@ size_t CBytestream::Read(NetworkSocket sock) {
 		if(res <= 0) break;
 		Data.append(buf, res);
 		len += res;
-		if(res < sizeof(buf)) break;
+		if((size_t)res < sizeof(buf)) break;
 	}
 	return len;
 }
